# contracts/ai-service.openapi.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "msisdn": {
            "anyOf": [
              {
                "pattern": "^\\d{10}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The customer open in the console, used as context when the question names no number.",
            "title": "Msisdn"
          },
          "query": {
            "description": "What the agent typed.",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "description": "The reply after moderation: safe to show to the agent.",
            "title": "Answer",
            "type": "string"
          },
          "dispute_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispute Id"
          },
          "pii_masked": {
            "title": "Pii Masked",
            "type": "boolean"
          },
          "route": {
            "description": "rag | balance | dispute | escalation | testgen",
            "title": "Route",
            "type": "string"
          },
          "route_method": {
            "description": "regex | llm: how the route was decided.",
            "title": "Route Method",
            "type": "string"
          },
          "safety_flag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Set when the safety check blocked the reply.",
            "title": "Safety Flag"
          },
          "sources": {
            "description": "Knowledge-base passages the answer was built from.",
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "test_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TestResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "ticket_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ticket Id"
          },
          "tool_calls": {
            "description": "Billing calls the agents made.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tool Calls",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "route",
          "route_method",
          "pii_masked"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ErrorEvent": {
        "description": "The `error` server-sent event: the request failed; `detail` is safe to show.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "ready": {
            "description": "False until the models have finished loading (the first minutes after start).",
            "title": "Ready",
            "type": "boolean"
          },
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "ready",
          "provider",
          "model"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "StepEvent": {
        "description": "One `step` server-sent event: an agent node finished.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "node": {
            "title": "Node",
            "type": "string"
          }
        },
        "required": [
          "node",
          "message"
        ],
        "title": "StepEvent",
        "type": "object"
      },
      "TestResult": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "evidence_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path to the screenshot, relative to this service.",
            "title": "Evidence Url"
          },
          "status": {
            "enum": [
              "pass",
              "fail"
            ],
            "title": "Status",
            "type": "string"
          },
          "steps_completed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Steps Completed"
          }
        },
        "required": [
          "status",
          "detail"
        ],
        "title": "TestResult",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Billing Ops AI service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chat": {
      "post": {
        "operationId": "chat_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat"
      }
    },
    "/chat/stream": {
      "post": {
        "operationId": "chat_stream_chat_stream_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Server-sent events. `step` (StepEvent) as each agent finishes, comment lines as keep-alives, then exactly one `result` (ChatResponse) or `error` (ErrorEvent), then `done`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream"
      }
    },
    "/evidence/{name}": {
      "get": {
        "operationId": "evidence_evidence__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}